backend/api/v1/routes/middleware.py
"""Custom middleware for API v1."""
from fastapi import Request
from starlette.middleware.base import BaseHTTPMiddleware


class AuditLogMiddleware(BaseHTTPMiddleware):
    """Middleware to log all requests for audit trail."""

    async def dispatch(self, request: Request, call_next):
        """Process request and log audit trail."""
        # TODO: Implement audit logging
        response = await call_next(request)
        return response
